+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v2.196.0 (2023-10-27)
+
+### Features
+
+ * inference instance type conditioned on training instance type
+
+### Bug Fixes and Other Changes
+
+ * improved jumpstart tagging
+
 ## v2.195.1 (2023-10-26)
 
 ### Bug Fixes and Other Changes